Review - split into isolated rendering file
